refactor: update language default to Arabic, enhance animations, and improve notification handling

+++ lib/feature/client/notification/logic/cubit/user_notification_cubit.dart
@@ -1,6 +1,5 @@
 import 'package:elminiawy/core/common/shared/shared_imports.dart';
 
-
 part 'user_notification_state.dart';
 part 'user_notification_cubit.freezed.dart';
 
@@ -8,43 +7,53 @@ class UserNotificationCubit extends Cubit<UserNotificationState> {
   List<UserNotificationData> dataList = [];
 
   UserNotificationCubit(this._repository)
-      : super(const UserNotificationState.initial());
+    : super(const UserNotificationState.initial());
 
   final UserNotificationRepositoryImplement _repository;
 
   Future<void> updateAllNotificationsToSeen() async {
+    if (isClosed) return;
+
     emit(const UserNotificationState.userNotificationLoading());
 
     final result = await _repository.updateAllUserNotificationToSeenRepo();
 
+    if (isClosed) return;
+
     result.when(
       success: (data) {
+        if (isClosed) return;
         dataList = [];
         dataList.addAll(data.data!);
         emit(UserNotificationState.userNotificationLoaded(data));
       },
       failure: (error) {
-        emit(UserNotificationState.userNotificationError(
-          error));
+        if (isClosed) return;
+        emit(UserNotificationState.userNotificationError(error));
       },
     );
   }
 
   Future<void> deleteUserNotification(String id) async {
+    if (isClosed) return;
+
     emit(const UserNotificationState.deleteUserNotificationLoading());
 
     final result = await _repository.deleteUserNotificationRepo(id);
 
+    if (isClosed) return;
+
     result.when(
       success: (data) {
+        if (isClosed) return;
         dataList = [];
         dataList.addAll(data.data!);
 
         emit(UserNotificationState.deleteUserNotificationLoaded(data));
       },
       failure: (error) {
-        emit(UserNotificationState.deleteUserNotificationError(
-           error));
+        if (isClosed) return;
+        emit(UserNotificationState.deleteUserNotificationError(error));
       },
     );
   }
